requirements.txt
Django==2.1.5
django-filter==2.0.0
djangorestframework==3.9.0
Markdown==3.0.1
pkg-resources==0.0.0
pytz==2018.7
